sqlite_to_postgres/requirements.txt
django==3.2
flake8==4.0
python_dotenv==1.0.0
psycopg2-binary
django-split-settings==1.2.0
python_dateutil==2.8.2
pytest==7.4.2
pydantic==2.3.0